Revert everything to float due to performance issues

+++ module/vision/FieldLineDetector/src/FieldLineDetector.cpp
@@ -29,7 +29,7 @@ namespace module::vision {
             // Use configuration here from file FieldLineDetector.yaml
             log_level = config["log_level"].as<NUClear::LogLevel>();
 
-            cfg.confidence_threshold = config["confidence_threshold"].as<double>();
+            cfg.confidence_threshold = config["confidence_threshold"].as<float>();
             cfg.cluster_points       = config["cluster_points"].as<int>();
         });
 
@@ -38,8 +38,8 @@ namespace module::vision {
             const auto& cls        = horizon.mesh->classifications;
             const auto& neighbours = horizon.mesh->neighbourhood;
             // Unit vectors from camera to a point in the mesh, in world space
-            const Eigen::Matrix<double, 3, Eigen::Dynamic>& uPCw = horizon.mesh->rays.cast<double>();
-            const int LINE_INDEX                                 = horizon.class_map.at("line");
+            const Eigen::Matrix<float, 3, Eigen::Dynamic>& uPCw = horizon.mesh->rays;
+            const int LINE_INDEX                                = horizon.class_map.at("line");
             // PARTITION INDICES AND CLUSTER
             // Get some indices to partition
             std::vector<int> indices(horizon.mesh->indices.size());
@@ -87,8 +87,8 @@ namespace module::vision {
                 for (const auto& idx : cluster) {
                     lines->points.push_back(uPCw.col(idx));
                     // Project the field line point onto the field plane
-                    Eigen::Vector3d rCWw = Eigen::Vector3d(Hwc.translation().x(), Hwc.translation().y(), 0.0);
-                    Eigen::Vector3d rPCw = uPCw.col(idx) * std::abs(Hwc.translation().z() / uPCw.col(idx).z()) + rCWw;
+                    Eigen::Vector3f rCWw = Eigen::Vector3f(Hwc.translation().x(), Hwc.translation().y(), 0.0);
+                    Eigen::Vector3f rPCw = uPCw.col(idx) * std::abs(Hwc.translation().z() / uPCw.col(idx).z()) + rCWw;
                     lines->rPCw.push_back(rPCw.head<2>());
                 }
             }
